pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "agora-python-server-sdk>=2.4.0",
    "pydub>=0.25.1",
]
